requirements.txt
streamlit==1.35.0
mysql-connector-python==8.3.0
openai>=1.0.0